requirements.txt
cent>=1.1.0
requests>=2.8.1
django-rest-messaging>=0.0.1